Replace BrowserRouter with newly introduced createBrowserRouter
createBrowserRouter is the recommended router for all React Router web projects. It also enables the v6.4 data APIs like loaders, actions, fetchers and more
