+++ frontend/src/index.js
@@ -4,7 +4,6 @@ import { PersistGate } from 'redux-persist/integration/react';
 import { Provider } from 'react-redux';
 import WebFont from 'webfontloader';
 import * as Sentry from '@sentry/react';
-import { BrowserRouter } from 'react-router-dom';
 import { BrowserTracing } from '@sentry/tracing';
 
 import App from './App';
@@ -32,9 +31,7 @@ ReactDOM.render(
   <Provider store={store}>
     <PersistGate loading={null} persistor={persistor}>
       <ConnectedIntl>
-        <BrowserRouter>
-          <App />
-        </BrowserRouter>
+        <App />
       </ConnectedIntl>
     </PersistGate>
   </Provider>,
